requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
fido2>=1.1.0
twilio>=8.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0
dj-database-url>=2.1.0